Refactor structinit into different files
